hubble, policy/correlation: populate policy log field based on log cookie
Based on the log cookie passed up by the datapath in the policy verdict
notification message, try to look up the corresponding log string and
use it to populate the policy log field in the Hubble flow.

Note that the lookup happens after policy correlation was performed (if
enabled) as the cookie allows a direct mapping to the corresponding
policy. This also works if hubble policy correlation is disabled as it
only consists of a single map lookup.

Signed-off-by: Tobias Klauser <tobias@cilium.io>