Update Gateway API dependency for Gateway API v1.4.0
Notably, this will cause us to _fail_ the Gateway API Conformance
tests, because we have not implemented GRPCRoute support for
GAMMA, and the tests do not have a way to say "I support Mesh,
but not GRPCRoute in Mesh".

Also changes `kubectl apply` in the Gateway API conformance
tests to be server-side. This is required for the v1.4.0
HTTPRoute CustomResourceDefinition object to apply, as
client-side apply creates a `last-applied-config` annotation
that is now greater than the maximum size of 256kb.

The Gateway API version bump also included a controller-tools
version bump, which required regeneration of the Cilium CRDs
as well.

Signed-off-by: Nick Young <nick@isovalent.com>