check-encryption-leak: minor log format improvements
This is just to improve the readability of the log output.
We could use a local variable to hold a value in case of overlay packets,
so that the inner is logged with `↳`, while in native mode it doesn't have it.
However, this would make the code less readable, and we might need this
space for further variables. This commit also inverts the orders of reported
fields for ICMP packets, displaying first Type and Code, then Flags, Fragmentation
Offset and ID.

Overlay packets:
[15:42:42:253484] [0xffff8c6fd0db8500] 172.18.0.3:54777 -> 172.18.0.4:8472 (proto: 17, encap: 1, skb: 0)
[15:42:42:253484] [0xffff8c6fd0db8500]  ↳ 10.244.2.87:0 -> 10.244.1.122:0 (len: 84, proto: 1, ifindex: 38, netns: f0000000, srcPod: 1 (internal: 0), dstPod: 1 (internal: 0), proxy: 0 (masqueraded: 0))
[15:42:42:253484] [0xffff8c6fd0db8500]   ↳ ICMP (Type: 8, Code: 0, IPFlags: .D., FragOff: 0, FragID: 47889)

Native:
[16:07:53:048992] [0xffff8c7043a7ad00]  ↳ 10.244.2.121:57638 -> 10.244.0.155:8080 (len: 52, proto: 6, ifindex: 38, netns: f0000000, srcPod: 1 (internal: 0), dstPod: 1 (internal: 0), proxy: 0 (masqueraded: 0))
[16:07:53:048992] [0xffff8c7043a7ad00]   ↳ TCP (TCPFlags: ...A...., Seq: 2471606865, Ack: 3837323371)

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>