Add support for GRPCRoute to GAMMA reconciler
This commit adds support for using GRPCRoutes for
east/west traffic routing with GAMMA.

It includes a small amount of refactoring of the
Gateway API reconciler as well, to enable the
HTTPRoute and GRPCRoute translation logic to share
code across both north/south and east/west
traffic paths more easily.

Unfortunately, I also needed to restructure the
GAMMA tests to use the same format as the north/south
Gateway API tests, which has made this a _large_
change.

The main changes:

* Gateway API conformance test workflow now requires
  `kubectl --server-side`, as the HTTPRoute CustomResourceDefinition
  will fail to apply without it (the `last-applied-config`
  annotation is too large to fit in 256kb). Updated `Makefile.kind`
  and `.github/workflows/conformance-gateway-api.yaml` accordingly.
* Added GRPCRoute support and small refactor to ingestion
  in `operator/pkg/mode/ingestion`, `gamma.go` and
  `gateway.go`.
* GAMMA reconciler in `operator/pkg/gateway-api` migrated
  to use `operator/pkg/gateway-api/indexers`, added in #41232.
* GRPCRoute support added to `operator/pkg/gateway-api` as well.
* Tests in `opreator/pkg/gateway-api/testdata/gamma` updated to
  use `input/` and `output/` directories, as in
  `operator/pkg/gateway-api/testdata/gateway-api`.

Signed-off-by: Nick Young <nick@isovalent.com>